refactor: result 타입을 int에서 long으로 변경

+++ src/main/java/calculator/CalculatorService.java
@@ -17,7 +17,7 @@ public class CalculatorService {
             outputHandler.guideComment();
             String userInput = inputHandler.getUserInput();
             Calculator calculator = new Calculator(userInput);
-            int result = calculator.calculate();  // 계산 수행
+            long result = calculator.calculate();  // 계산 수행
             outputHandler.printSum(result); // 결과 출력
         } catch (IllegalArgumentException e) {
             throw new IllegalArgumentException(e.getMessage());
